+++ src/interleave/__init__.py
@@ -33,7 +33,7 @@ from typing import (
     cast,
 )
 
-__version__ = "0.2.0.dev1"
+__version__ = "0.2.0"
 __author__ = "John Thorvald Wodder II"
 __author_email__ = "interleave@varonathe.org"
 __license__ = "MIT"